Added some more data models Started an API Controller for Points Spent a LOT of time trying to get the javascipt to call the API successfully and went deep down the Promises rabbit hole. haven't checked that bit in though :)

+++ gucci/source/gucci-frontend/Models/Point.cs
@@ -5,7 +5,7 @@ namespace gucci_frontend.Models
     public class Point
     {
         public int id { get; set; }
-        public float longtitude { get; set; }
-        public float latitude { get; set; }
+        public double longtitude { get; set; }
+        public double latitude { get; set; }
     }
 }
